+++ staging/src/k8s.io/client-go/discovery/BUILD
@@ -14,7 +14,6 @@ go_library(
         "doc.go",
         "helper.go",
         "round_tripper.go",
-        "unstructured.go",
     ],
     importmap = "k8s.io/kubernetes/vendor/k8s.io/client-go/discovery",
     importpath = "k8s.io/client-go/discovery",
